moving ResultsValueType to model and modify imports
