+++ agent/src/main/scala/za/co/absa/atum/agent/model/MeasureResult.scala
@@ -17,7 +17,7 @@
 package za.co.absa.atum.agent.model
 
 import za.co.absa.atum.agent.exception.AtumAgentException.MeasurementException
-import za.co.absa.atum.model.utils.ResultValueType
+import za.co.absa.atum.model.ResultValueType
 
 /**
  *  This trait defines a contract for a measure result.
