+++ agent/src/test/scala/za/co/absa/atum/agent/model/MeasureUnitTests.scala
@@ -23,7 +23,7 @@ import za.co.absa.atum.agent.AtumContext.AtumPartitions
 import za.co.absa.atum.agent.model.AtumMeasure.{AbsSumOfValuesOfColumn, RecordCount, SumOfHashesOfColumn, SumOfValuesOfColumn}
 import za.co.absa.spark.commons.test.SparkTestBase
 import za.co.absa.atum.agent.AtumContext._
-import za.co.absa.atum.model.utils.ResultValueType
+import za.co.absa.atum.model.ResultValueType
 
 class MeasureUnitTests extends AnyFlatSpec with Matchers with SparkTestBase { self =>
