+++ server/src/test/scala/za/co/absa/atum/server/api/database/runs/functions/WriteCheckpointIntegrationTests.scala
@@ -16,9 +16,9 @@
 
 package za.co.absa.atum.server.api.database.runs.functions
 
+import za.co.absa.atum.model.ResultValueType
 import za.co.absa.atum.model.dto.{CheckpointDTO, MeasureDTO, MeasureResultDTO, MeasurementDTO, PartitionDTO}
 import za.co.absa.atum.model.dto.MeasureResultDTO.TypedValue
-import za.co.absa.atum.model.utils.ResultValueType
 import za.co.absa.atum.server.ConfigProviderTest
 import za.co.absa.atum.server.api.TestTransactorProvider
 import za.co.absa.atum.server.api.database.PostgresDatabaseProvider
